+++ docker/core/mkwebaxum/src/mk_lib_database_media_remote.rs
@@ -11,7 +11,7 @@ use stdext::function_name;
 
 pub async fn mk_lib_database_remote_media_count(
     sqlx_pool: &sqlx::PgPool,
-) -> Result<i32, sqlx::Error> {
+) -> Result<i64, sqlx::Error> {
     #[cfg(debug_assertions)]
     {
         mk_lib_logging::mk_logging_post_elk(
@@ -21,7 +21,7 @@ pub async fn mk_lib_database_remote_media_count(
         .await
         .unwrap();
     }
-    let row: (i32,) = sqlx::query_as("select count(*) from mm_media_remote")
+    let row: (i64,) = sqlx::query_as("select count(*) from mm_media_remote")
         .fetch_one(sqlx_pool)
         .await?;
     Ok(row.0)
